% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{betaABfromMeanKappa}
\alias{betaABfromMeanKappa}
\title{Shape parameters from central tendency and scale:}
\usage{
betaABfromMeanKappa(mean, kappa)
}
\arguments{
\item{mean}{mean}

\item{kappa}{kappa}
}
\value{
betaABfromMeanKappa
}
\description{
Shape parameters from central tendency and scale:
}
