% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{betaABfromMeanSD}
\alias{betaABfromMeanSD}
\title{Shape parameters from central tendency and scale:}
\usage{
betaABfromMeanSD(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{sd}
}
\value{
betaABfromMeanSD
}
\description{
Shape parameters from central tendency and scale:
}
