% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{betaABfromModeKappa}
\alias{betaABfromModeKappa}
\title{Title}
\usage{
betaABfromModeKappa(mode, kappa)
}
\arguments{
\item{mode}{mode}

\item{kappa}{kappa}
}
\value{
betaABfromModeKappa
}
\description{
Title
}
