% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{DbdaAcfPlot}
\alias{DbdaAcfPlot}
\title{plotting properties of mcmc coda objects.}
\usage{
DbdaAcfPlot(codaObject, parName = varnames(codaObject)[1], plColors = NULL)
}
\arguments{
\item{codaObject}{codaObject}

\item{parName}{parName}

\item{plColors}{plColors}
}
\description{
plotting properties of mcmc coda objects.
}
