% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{diagMCMC}
\alias{diagMCMC}
\title{plotting properties of mcmc coda objects.}
\usage{
diagMCMC(codaObject, parName = varnames(codaObject)[1], saveName = NULL,
  saveType = "jpg")
}
\arguments{
\item{codaObject}{codaObject}

\item{parName}{parName}

\item{saveName}{saveName}

\item{saveType}{saveType}
}
\description{
plotting properties of mcmc coda objects.
}
