% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{diagStanFit}
\alias{diagStanFit}
\title{plotting properties of mcmc coda objects.}
\usage{
diagStanFit(stanFit, parName, saveName = NULL, saveType = "jpg")
}
\arguments{
\item{stanFit}{stanFit}

\item{parName}{parName}

\item{saveName}{saveName}

\item{saveType}{saveType}
}
\value{
diagStanFit
}
\description{
plotting properties of mcmc coda objects.
}
