% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{gammaShRaFromMeanSD}
\alias{gammaShRaFromMeanSD}
\title{Shape parameters from central tendency and scale:}
\usage{
gammaShRaFromMeanSD(mean, sd)
}
\arguments{
\item{mean}{mean}

\item{sd}{sd}
}
\value{
gammaShRaFromMeanSD
}
\description{
Shape parameters from central tendency and scale:
}
