% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{gammaShRaFromModeSD}
\alias{gammaShRaFromModeSD}
\title{Shape parameters from central tendency and scale:}
\usage{
gammaShRaFromModeSD(mode, sd)
}
\arguments{
\item{mode}{mode}

\item{sd}{sd}
}
\value{
gammaShRaFromModeSD
}
\description{
Shape parameters from central tendency and scale:
}
