% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{HDIofGrid}
\alias{HDIofGrid}
\title{Title}
\usage{
HDIofGrid(probMassVec, credMass = 0.95)
}
\arguments{
\item{probMassVec}{a vector of probability masses at each grid point.}

\item{credMass}{the desired mass of the HDI region.}
}
\value{
A list with components: indices is a vector of indices that are in the HDI: mass is the total mass of the included indices :height is the smallest component probability mass in the HDI
}
\description{
Title
}
\examples{
#For determining HDI of a beta(30,12) distribution approximated on a grid:
 probDensityVec = dbeta( seq(0,1,length=201) , 30 , 12 )
 probMassVec = probDensityVec / sum( probDensityVec )
 HDIinfo = HDIofGrid( probMassVec )
 show( HDIinfo )

}
