% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{HDIofICDF}
\alias{HDIofICDF}
\title{Title}
\usage{
HDIofICDF(ICDFname, credMass = 0.95, tol = 1e-08, ...)
}
\arguments{
\item{ICDFname}{R's name for the inverse cumulative density function of the distribution.}

\item{credMass}{the desired mass of the HDI region.}

\item{tol}{tol is passed to R's optimize function.}
}
\value{
Highest density iterval (HDI) limits in a vector.
}
\description{
Title
}
\examples{
#For determining HDI of a beta(30,12) distribution, type
  HDIofICDF( qbeta , shape1 = 30 , shape2 = 12 )
  #Notice that the parameters of the ICDFname must be explicitly named;
  #e.g., HDIofICDF( qbeta , 30 , 12 ) does not work.
#Adapted and corrected from Greg Snow's TeachingDemos package.
}
