% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{HDIofMCMC}
\alias{HDIofMCMC}
\title{Functions for computing limits of HDI's: Computes highest density interval from a sample of representative values, estimated as shortest credible interval.}
\usage{
HDIofMCMC(sampleVec, credMass = 0.95)
}
\arguments{
\item{sampleVec}{a vector of representative values from a probability distribution.}

\item{credMass}{is a scalar between 0 and 1, indicating the mass within the credible interval that is to be estimated.}
}
\value{
a vector containing the limits of the HDI
}
\description{
Functions for computing limits of HDI's: Computes highest density interval from a sample of representative values, estimated as shortest credible interval.
}
