% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{openGraph}
\alias{openGraph}
\title{Function for opening graphics that operate the same for
Windows and Macintosh and Linux operating systems. At least, that's the hope!}
\usage{
openGraph(width = 7, height = 7, mag = 1, ...)
}
\arguments{
\item{width}{width}

\item{height}{height}

\item{mag}{mag}
}
\description{
Function for opening graphics that operate the same for
Windows and Macintosh and Linux operating systems. At least, that's the hope!
}
