% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{plotPost}
\alias{plotPost}
\title{Title}
\usage{
plotPost(paramSampleVec, cenTend = c("mode", "median", "mean")[1],
  compVal = NULL, ROPE = NULL, credMass = 0.95, HDItextPlace = 0.7,
  xlab = NULL, xlim = NULL, yaxt = NULL, ylab = NULL, main = NULL,
  cex = NULL, cex.lab = NULL, col = NULL, border = NULL,
  showCurve = FALSE, breaks = NULL, ...)
}
\arguments{
\item{paramSampleVec}{paramSampleVec}

\item{cenTend}{cenTend}

\item{compVal}{compVal}

\item{ROPE}{ROPE}

\item{credMass}{credMass}

\item{HDItextPlace}{HDItextPlace}

\item{xlab}{xlab}

\item{xlim}{xlim}

\item{yaxt}{yaxt}

\item{ylab}{ylab}

\item{main}{main}

\item{cex}{cex}

\item{cex.lab}{cex.lab}

\item{col}{col}

\item{border}{border}

\item{showCurve}{showCurve}

\item{breaks}{breaks}
}
\description{
Title
}
