% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{saveGraph}
\alias{saveGraph}
\title{Function for saving graphics that operate the same for
Windows and Macintosh and Linux operating systems. At least, that's the hope!}
\usage{
saveGraph(file = "saveGraphOutput", type = "pdf", ...)
}
\arguments{
\item{file}{file}

\item{type}{type}
}
\description{
Function for saving graphics that operate the same for
Windows and Macintosh and Linux operating systems. At least, that's the hope!
}
