% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBDA2E.R
\name{summarizePost}
\alias{summarizePost}
\title{Title}
\usage{
summarizePost(paramSampleVec, compVal = NULL, ROPE = NULL,
  credMass = 0.95)
}
\arguments{
\item{paramSampleVec}{paramSampleVec}

\item{compVal}{compVal}

\item{ROPE}{ROPE}

\item{credMass}{credMass}
}
\value{
summarizePost
}
\description{
Title
}
